requirements.txt
openai
anthropic
python-dotenv
typing_extensions
